Renamed create-search-stream module to lower kebab case.
