+++ server/index.js
@@ -5,7 +5,7 @@ const socketio = require('socket.io');
 const errorMiddleware = require('./error-middleware');
 const staticMiddleware = require('./static-middleware');
 const authorizationMiddleware = require('./authorization-middleware');
-const { createSearchStream, parseKeywords } = require('./createSearchStream');
+const { createSearchStream, parseKeywords } = require('./create-search-stream');
 const jwt = require('jsonwebtoken');
 const cookieParser = require('cookie-parser');
 const Snoowrap = require('snoowrap');
